fix: project area unit translations and refactor useOrders
